시간 경과 계산 함수 분리
